+++ presto-orc/src/main/java/com/facebook/presto/orc/reader/LongStreamReader.java
@@ -173,7 +173,7 @@ public class LongStreamReader
     }
 
     @Override
-    public void setResultSizeBudget(int bytes)
+    public void setResultSizeBudget(long bytes)
     {
         currentReader.setResultSizeBudget(bytes);
     }
